Finalização do carrinho shop

+++ Pet_Shop/wwwroot/js/controllers/ProdutoController.js
@@ -30,7 +30,7 @@ function BuscaProduto(codProduto) {
     return fetch(url) 
         .then(response => {
             if (!response.ok) {
-                throw new Error('Erro ao buscar produto');
+                MessageBox('Atenção', 'Erro ao carregar produto');
             }
             return response.json();
         })
@@ -39,8 +39,7 @@ function BuscaProduto(codProduto) {
             return produto; 
         })
         .catch(error => {
-            console.error('Erro ao buscar produto:', error);
-            throw error; 
+            MessageBox('Atenção', 'Erro ao carregar produto');
         });
 }
 
@@ -90,13 +89,20 @@ export function AddCarrinho() {
 
 function SaveLocalStorage(produto) {
     let carrinho = JSON.parse(localStorage.getItem('carrinho')) || [];//carrega carrinho
+    let exist = false;
     carrinho.forEach(item => {
         if (item._cod == produto.cod) {
-            MessageBox('Atenção', 'O produto já foi Incluido no Carrinho.')
+            exist = true;
         }
     });
-    carrinho.push(produto);//add produto no array
-    localStorage.setItem('carrinho', JSON.stringify(carrinho));//set novo carrinho com produto novo
+
+    if (exist == false) {
+        carrinho.push(produto);//add produto no array
+        localStorage.setItem('carrinho', JSON.stringify(carrinho));//set novo carrinho com produto novo
+    }
+    else {
+        MessageBox('Atenção', 'O produto já foi Incluido no Carrinho.')
+    }
 }
 
 export function LoadCarrinho() {
@@ -104,21 +110,33 @@ export function LoadCarrinho() {
     document.getElementById('overlay').style.display = 'block';
     var itemsCarrinho = JSON.parse(localStorage.getItem('carrinho')) || [];
     const itemModelo = document.querySelector('.item-carrinho');
+    let totalCarrinho = 0;
 
-    itemsCarrinho.forEach(produto => {
-        const novoItem = itemModelo.cloneNode(true);//clona
+    if (itemsCarrinho.length === 0) {
+        document.getElementById('valor-carrinho').textContent = "Carrinho Vazio";
+        document.getElementById('continuar-pedido').style.display = 'none';
+    }
+    else {
+        document.getElementById('continuar-pedido').style.display = 'flex';
+        itemsCarrinho.forEach(produto => {
+            const novoItem = itemModelo.cloneNode(true);//clona
 
-        novoItem.style.display = 'Flex';
-        novoItem.setAttribute('class', 'card clone-item');
-        novoItem.querySelector('.card-img-top').setAttribute('src', produto._imagem);
-        novoItem.querySelector('.btn-remove-carrinho').setAttribute('id', `btn-remove-${produto._cod}`);
-        novoItem.querySelector('.card-text:nth-of-type(1)').textContent = produto._nome;
-        novoItem.querySelector('.card-text:nth-of-type(2)').textContent = produto._descricao;
-        novoItem.querySelector('.row-text-item:nth-of-type(2) .card-text:nth-of-type(1)').textContent = produto._valor;
-        novoItem.querySelector('.row-text-item:nth-of-type(2) .card-text:nth-of-type(2)').textContent = "Quantidade:"+produto._quantidade;
+            novoItem.style.display = 'Flex';
+            novoItem.setAttribute('class', 'card clone-item');
+            novoItem.querySelector('.card-img-top').setAttribute('src', produto._imagem);
+            novoItem.querySelector('.btn-remove-carrinho').setAttribute('id', `btn-remove-${produto._cod}`);
+            novoItem.querySelector('.card-text:nth-of-type(1)').textContent = produto._nome;
+            novoItem.querySelector('.card-text:nth-of-type(2)').textContent = produto._descricao;
+            novoItem.querySelector('.row-text-item:nth-of-type(2) .card-text:nth-of-type(1)').textContent = produto._valor;
+            novoItem.querySelector('.row-text-item:nth-of-type(2) .card-text:nth-of-type(2)').textContent = "Quantidade:" + produto._quantidade;
 
-        document.querySelector('.carrinho-items').appendChild(novoItem);//insere no html
-    });
+            document.querySelector('.carrinho-items').appendChild(novoItem);//insere no html
+
+            totalCarrinho = totalCarrinho + parseInt((produto._valor).substring(2));
+        });
+
+        document.getElementById('valor-carrinho').textContent = "R$ " + totalCarrinho + ".00";
+    }
 }
 
 export function RemoveLocalStorage(codRemove) {
@@ -133,26 +151,19 @@ export function RemoveLocalStorage(codRemove) {
 
         LoadCarrinho();
     } else {
-        console.error('Código inválido:', codRemove);
+        MessageBox('Atenção', 'Erro ao remover produto.');
     }
 }
 
-export function FecharCarrinho() {//erro ao abrir, fechar e abrir de novo
+export function FecharCarrinho() { 
     document.getElementById('carrinho').style.display = 'none'
     document.getElementById('overlay').style.display = 'none';
     document.querySelector('.clone-item').remove();
 }
 
-function MessageBox(titulo, mensagem) {;//colocar tratamento de exeção
-    const url = `MessageBox?titulo=${titulo}&mensagem=${mensagem}`;
-    fetch(url)
-        .then(response => response.json())
-        .then(data => {
-
-        })
-        .catch(error => console.error('Erro:', error));
-
-    window.location.href = "/Shared/MessageBox.cshtml";
+function MessageBox(titulo, mensagem) {
+    const url = `MessageBox?titulo=${encodeURIComponent(titulo)}&mensagem=${encodeURIComponent(mensagem)}`;
+    window.location.href = url;
 }
